services/user-service/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0

